requirements.txt
# Dependencias generales del portafolio ClinBioinfo
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
biopython==1.81
streamlit==1.25.0
plotly==5.15.0
jupyter==1.0.0
sqlite3